# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RankRight API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentUser {
  userId: string;
  email: string;
  name?: string | null;
  agencyId: string;
  agencyName: string;
}

export interface Client {
  id: string;
  name: string;
  websiteUrl?: string | null;
  city?: string | null;
  phone?: string | null;
  placeId?: string | null;
  createdAt: string;
}

export interface CreateClientInput {
  /** @minLength 1 */
  name: string;
  websiteUrl?: string | null;
  city?: string | null;
  phone?: string | null;
}

export interface RunAuditInput {
  clientId: string;
}

export type AuditStatus = typeof AuditStatus[keyof typeof AuditStatus];


export const AuditStatus = {
  running: 'running',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface Audit {
  id: string;
  clientId: string;
  status: AuditStatus;
  score?: number | null;
  startedAt: string;
  completedAt?: string | null;
  error?: string | null;
}

export type AuditCheckCategory = typeof AuditCheckCategory[keyof typeof AuditCheckCategory];


export const AuditCheckCategory = {
  gbp: 'gbp',
  seo: 'seo',
  performance: 'performance',
  nap: 'nap',
  tech: 'tech',
  analytics: 'analytics',
} as const;

export type AuditCheckStatus = typeof AuditCheckStatus[keyof typeof AuditCheckStatus];


export const AuditCheckStatus = {
  pass: 'pass',
  warn: 'warn',
  fail: 'fail',
  stub: 'stub',
  error: 'error',
} as const;

export interface AuditCheck {
  id: string;
  checkKey: string;
  title: string;
  category: AuditCheckCategory;
  status: AuditCheckStatus;
  score: number;
  weight: number;
  finding: string;
  recommendation: string;
}

export interface AuditDetail {
  audit: Audit;
  client: Client;
  checks: AuditCheck[];
}