# lib/api-zod/src/generated/types/auditDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RankRight API
 * OpenAPI spec version: 0.1.0
 */
import type { Audit } from './audit';
import type { AuditCheck } from './auditCheck';
import type { Client } from './client';

export interface AuditDetail {
  audit: Audit;
  client: Client;
  checks: AuditCheck[];
}