# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RankRight API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: string;
  name: string;
  websiteUrl?: string | null;
  city?: string | null;
  phone?: string | null;
  placeId?: string | null;
  createdAt: Date;
}