# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RankRight API
 * OpenAPI spec version: 0.1.0
 */

export interface CurrentUser {
  userId: string;
  email: string;
  name?: string | null;
  agencyId: string;
  agencyName: string;
}