# lib/db/src/schema/clients.ts
import { pgTable, text, timestamp, uuid, index } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { agenciesTable } from "./agencies";

export const clientsTable = pgTable(
  "clients",
  {
    id: uuid("id").primaryKey().defaultRandom(),
    agencyId: uuid("agency_id")
      .notNull()
      .references(() => agenciesTable.id, { onDelete: "cascade" }),
    name: text("name").notNull(),
    websiteUrl: text("website_url"),
    city: text("city"),
    placeId: text("place_id"),
    phone: text("phone"),
    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  },
  (t) => [index("clients_agency_id_idx").on(t.agencyId)],
);

export const insertClientSchema = createInsertSchema(clientsTable, {
  websiteUrl: z.string().url().nullable().optional(),
}).omit({
  id: true,
  agencyId: true,
  createdAt: true,
  placeId: true,
});
export type InsertClient = z.infer<typeof insertClientSchema>;
export type Client = typeof clientsTable.$inferSelect;